Moving go client rows and streaming_rows out of vtgateconn and into the go client driver.
